more conversion warnings
unsigned is hardcoded in boost

+++ Polygon_mesh_processing/include/CGAL/Polygon_mesh_processing/connected_components.h
@@ -387,7 +387,7 @@ std::size_t keep_largest_connected_components(PolygonMesh& pmesh,
   Output_iterator out = choose_parameter<Output_iterator>(get_parameter(np, internal_np::output_iterator));
 
   // vector_property_map
-  boost::vector_property_map<std::size_t, FaceIndexMap> face_cc(num_faces(pmesh), fimap);
+  boost::vector_property_map<std::size_t, FaceIndexMap> face_cc(static_cast<unsigned>(num_faces(pmesh)), fimap);
   std::size_t num = connected_components(pmesh, face_cc, np);
 
   // Even if we do not want to keep anything we need to first
@@ -518,7 +518,7 @@ std::size_t keep_large_connected_components(PolygonMesh& pmesh,
   Output_iterator out = choose_parameter<Output_iterator>(get_parameter(np, internal_np::output_iterator));
 
   // vector_property_map
-  boost::vector_property_map<std::size_t, FaceIndexMap> face_cc(num_faces(pmesh), fim);
+  boost::vector_property_map<std::size_t, FaceIndexMap> face_cc(static_cast<unsigned>(num_faces(pmesh)), fim);
   std::size_t num = connected_components(pmesh, face_cc, np);
   std::vector<Face_size> component_size(num, 0);
 
@@ -593,7 +593,7 @@ void keep_or_remove_connected_components(PolygonMesh& pmesh
   for(std::size_t i : components_to_keep)
     cc_to_keep.insert(i);
 
-  boost::vector_property_map<bool, VertexIndexMap> keep_vertex(num_vertices(pmesh), vim);
+  boost::vector_property_map<bool, VertexIndexMap> keep_vertex(static_cast<unsigned>(num_vertices(pmesh)), vim);
   for(vertex_descriptor v : vertices(pmesh))
     keep_vertex[v] = false;
 
@@ -831,7 +831,7 @@ void remove_connected_components(PolygonMesh& pmesh
   typedef typename CGAL::GetInitializedFaceIndexMap<PolygonMesh, CGAL_PMP_NP_CLASS>::type FaceIndexMap;
   FaceIndexMap fim = CGAL::get_initialized_face_index_map(pmesh, np);
 
-  boost::vector_property_map<std::size_t, FaceIndexMap> face_cc(num_faces(pmesh), fim);
+  boost::vector_property_map<std::size_t, FaceIndexMap> face_cc(static_cast<unsigned>(num_faces(pmesh)), fim);
   connected_components(pmesh, face_cc, np);
 
   std::vector<std::size_t> cc_to_remove;
@@ -881,7 +881,7 @@ void keep_connected_components(PolygonMesh& pmesh
   typedef typename CGAL::GetInitializedFaceIndexMap<PolygonMesh, CGAL_PMP_NP_CLASS>::type FaceIndexMap;
   FaceIndexMap fim = CGAL::get_initialized_face_index_map(pmesh, np);
 
-  boost::vector_property_map<std::size_t, FaceIndexMap> face_cc(num_faces(pmesh), fim);
+  boost::vector_property_map<std::size_t, FaceIndexMap> face_cc(static_cast<unsigned>(num_faces(pmesh)), fim);
   connected_components(pmesh, face_cc, np);
 
   std::vector<std::size_t> cc_to_keep;
